chore(qwen3-scheduler): prune redundant fake-loop tests, extract test module
scheduler.rs's in-file test module had grown to ~1555 lines — larger than
the code it tested — by accumulating fake-executor-driven loop tests that
either duplicate the real-GPU e2e in tests/ or assert on mock spy recordings.

Drop 11 low-value tests (27 -> 16):
- 5 fake-loop tests covered by tests/{scheduler_robustness,context_window,
  cached_tokens_usage}.rs: consumer/receiver drop, oversized-echo & over-
  context rejection, cached-tokens-once, long-prompt chunk completion
- 3 reject-through-loop tests whose decision is already covered by the pure
  admit_deferred_requests / lifetime-block unit tests
- mixed_lora_..._one_batch: asserted on Arc<Mutex<Vec>> spy recordings of
  internal batch grouping (testing the mock, not behaviour)
- active_receiver_drop (dup of scheduler_robustness e2e) and
  lora_control_reports_unimplemented (asserts a stub is a stub)

Drop the now-unused FakeExecutor batch-recorder fields/builders. The surviving
harness only does fault injection (decode failure), missing-adapter rejection
(no e2e covers it), control-channel idle-gating, and side-effect observation
(KV release / echo-not-prefetched) — things the real-GPU e2e can't reproduce.

Move the remaining #[cfg(test)] module to src/scheduler/tests.rs (flat layout,
`use super::*` preserves crate-private access). scheduler.rs: 2672 -> 1119.

Kept: 8 pure functional-core tests + 2 BlockPool peak-draw property tests +
the 6 fault/control/side-effect tests above.
